Namespace Refactoring: Transition from 'Elements' to 'Views' for PDF-visible objects (Related to issue #78) - Removed outdated files and their associated renderers. - Updated imports and references across builders, interfaces, tests, and other relevant files to adopt 'Views' in place of 'Elements'. - Introduced new Views directories to support the updated structure. - Applied minor adjustments to ensure full alignment with the revised namespace.
