+++ MauiPdfGenerator/Fluent/Interfaces/Views/IPdfImage.cs
@@ -1,4 +1,4 @@
-﻿namespace MauiPdfGenerator.Fluent.Interfaces.Elements;
+namespace MauiPdfGenerator.Fluent.Interfaces.Views;
 
 public interface IPdfImage<TSelf> : IPdfElement<TSelf> where TSelf : IPdfElement<TSelf>
 {
